Moved helper functions into utils. _check_event can't be easily moved atm because it leads to cirulate includings ...
